requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
alembic==1.12.1
pydantic==2.5.2
firebase-admin==6.2.0
yoomoney==2.4.0
google-maps-services-python==3.3.1
websockets==12.0
redis==5.0.1 